+++ NEWS.md
@@ -2,7 +2,7 @@
 
 * Add tidiers for compatability with `broom`.
 
-* Add `h0` argument for specifying the null hypothesis (replaces the previous 
-`pvals` argument and takes on different functionality)
+* Bundle support for different null hypotheses as part of the `resampvar` string
+argument (replaces and extends the previous `pvals` and `h0` arguments).
 
 * Initial GitHub release.